nlx-createsurvey/requirements.txt
pymongo
kafka-python
webbrowser
pygame 
ftplib
pyttsx3 
speechRecognition 
pyaudio
wave
